abstract application logic into separate component
